requirements/base.txt
Django==1.11.1
pytz==2017.2
django-filter==1.0.2
djangorestframework==3.6.2
Markdown==2.6.8
